django-furniture_factory/furniture_factory/requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.17
Pygments==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
